chore: add remove domain

+++ apps/admin/src/graphql/services/domain/removeDomain.ts
@@ -15,7 +15,34 @@ export const removeDomain = async (
   const domain = await prisma.domain.findFirst({
     where: { author_id: session.user.id },
   });
-  await prisma.domain.delete({ where: { author_id: session.user.id } });
 
-  return { ok: true, message: "Domain removed" };
+  if (!domain) {
+    return {
+      ok: false,
+      message: "No domain found",
+    };
+  }
+  try {
+    const response = await fetch(
+      `https://api.vercel.com/v9/projects/${process.env.VERCEL_PROJECT_ID}/domains/${domain.name}?teamId=${process.env.VERCEL_TEAM_ID}`,
+      {
+        headers: {
+          Authorization: `Bearer ${process.env.VERCEL_AUTH_BEARER_TOKEN}`,
+        },
+        method: "DELETE",
+      }
+    );
+
+    const json = await response.json();
+    // eslint-disable-next-line no-console
+    console.log(json);
+    await prisma.domain.delete({ where: { author_id: session.user.id } });
+
+    return { ok: true, message: "Domain removed" };
+  } catch (err: any) {
+    return {
+      ok: false,
+      message: err.message,
+    };
+  }
 };
